+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatNativeWindowBorder.java
@@ -225,6 +225,10 @@ public class FlatNativeWindowBorder
 		if( !SystemInfo.isWindows_10_orLater )
 			return;
 
+		// do not use when running in JetBrains Projector
+		if( SystemInfo.isProjector )
+			return;
+
 		// check whether disabled via system property
 		if( !FlatSystemProperties.getBoolean( FlatSystemProperties.USE_WINDOW_DECORATIONS, true ) )
 			return;